rename template file
